requirements.txt
python-docx>=1.1.0
pyyaml>=6.0
lxml>=4.9
flask>=3.0
pillow>=10.0
